layoutSubjectsを移植
